requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
python-multipart==0.0.6
PyPDF2==3.0.1
python-docx==0.8.11
ftfy==6.1.1
transformers>=4.37.0
accelerate>=0.26.0
sentence-transformers==2.2.2
faiss-cpu==1.7.4
langchain==0.0.267
datasets==2.14.4
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0
spacy==3.6.1
numpy>=1.20.0
packaging 
bitsandbytes
accelerate
